+++ src/org/apache/mina/tcp/base/transserver/protocol/connect/TServerDSReceiveConnectWriter.java
@@ -19,7 +19,7 @@ public class TServerDSReceiveConnectWriter extends TCPBaseWriter
     }
     
     @Override
-    public int GetSrcServerId()
+    public long GetSrcServerId()
     {
     	return TServerConfig.SERVER_ID;
     }
