+++ src/org/apache/mina/tcp/base/transserver/protocol/transmission/TServerToClientTransWriter.java
@@ -14,7 +14,7 @@ public class TServerToClientTransWriter extends TCPBaseWriter
     }
     
     @Override
-    public int GetSrcServerId()
+    public long GetSrcServerId()
     {
     	return TServerConfig.SERVER_ID;
     }
