+++ src/org/apache/mina/tcp/base/transserver/protocol/transmission/TServerToTServerTransWriter.java
@@ -22,7 +22,7 @@ public class TServerToTServerTransWriter extends TCPBaseWriter
     	this.datas          = datas;
     }
     
-    public int GetSrcServerId()
+    public long GetSrcServerId()
     {
     	return TServerConfig.SERVER_ID;
     }
